requirements.txt
Django
djangorestframework
django-environ
django-cors-headers
django-extensions
psycopg2
djangorestframework-simplejwt
coreapi
PyYAML
django-notifications-hq
channels
channels-redis
asgiref
daphne
twisted[http2,tls]
python-decouple
whitenoise